DQM changes - run_list is wiped when a new run appears in case of a dqm machine.

+++ python/hltd.py
@@ -962,7 +962,8 @@ class Run:
                         logging.info('process '+str(resource.process.pid)+' terminated')
                     logging.info('releasing resource(s) '+str(resource.cpu))
                     for cpu in resource.cpu:
-                      os.rename(used+cpu,idles+cpu)
+                      if os.path.exists(used+cpu):
+                        os.rename(used+cpu,idles+cpu)
                     resource.process=None
 
             self.online_resource_list = []
@@ -1196,6 +1197,8 @@ class RunRanger:
                     if conf.dqm_machine:
                         for run in run_list:
                             run.Shutdown()
+
+                        run_list = []
 
                     run_list.append(Run(nr,event.fullpath,bu_dir))
                     resource_lock.acquire()
